Fix warnings when compiling with MSVC 2005 at warning level 3
darcs-hash:20060824165414-11237-cacab8f80a76d322cdf18d23cef1e38a6de4d15a.gz

+++ libsofia-sip-ua/nua/nua_event_server.c
@@ -257,7 +257,7 @@ void nua_stack_authorize(nua_t *nua,
 
   tl_gets(tags,
 	  NEATAG_SUB_REF(sub),
-	  NUTAG_SUBSTATE_REF(state),
+	  NUTAG_SUBSTATE_REF((int)state),
 	  TAG_END());
 
   if (sub && state > 0) {
